pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sourcerer"
version = "1.0.0"
description = "AI-powered content aggregation and generation system"
authors = [{name = "Sourcerer Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.2",
    "cryptography>=41.0.8",
    "pyyaml>=6.0.1",
    "apscheduler>=3.10.4",
    "filelock>=3.13.1",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.2",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.4",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "openai>=1.3.7",
    "anthropic>=0.7.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
sourcerer = "backend.main:cli"

[tool.setuptools]
packages = ["backend", "frontend"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"